% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet-package.R
\docType{package}
\name{daymet-package}
\alias{daymet-package}
\alias{daymet}
\title{daymet-package}
\description{
Provides simplified access to OPeNDAP [daymet](https://daymet.ornl.gov/) data served by ORNL thredds server.
}
