% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{daymet_bag}
\alias{daymet_bag}
\title{Construct a daymet ncdf object bag of goodies}
\usage{
daymet_bag(x, verbose = TRUE)
}
\arguments{
\item{x}{a ncdf4 object}

\item{verbose}{logical if TRUE then issue a polite message about the wait}
}
\value{
bag a list (bag of goodies) useful for navigating the ncdf object.
It includes the following elements
\itemize{
    \item{param character, the parameter name}
    \item{crs character the CRS projection string}
    \item{res numeric the grid cell size in meters}
    \item{lon numeric matrix of cell locations in degrees longitude}
    \item{lat numeric matrix of cell locations in degrees latitude}
    \item{x the east-west cell locations (meters)}
    \item{y the north south cell locations (meters)}
}
}
\description{
Construct a daymet ncdf object bag of goodies
}
