% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{daymet_grid_uri}
\alias{daymet_grid_uri}
\title{Build a daymet uri for gridded data}
\usage{
daymet_grid_uri(
  year = 2018,
  mosaic = "na",
  param = "tmin",
  interval = "daily",
  baseuri = "https://thredds.daac.ornl.gov/thredds/dodsC/ornldaac"
)
}
\arguments{
\item{year}{4 digit year (just one please)}

\item{mosaic}{character mosaic set (na, pr or hi)}

\item{param}{character the parameter name}

\item{interval}{character daily, monthly or annual}

\item{baseuri}{character, the base uri}
}
\value{
charcater uri
}
\description{
Build a daymet uri for gridded data
}
