% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{daymet_index}
\alias{daymet_index}
\title{Compute indices into daymet arrays from lon/lat or from the native
daymet Lambert conformal conic coordinates}
\usage{
daymet_index(lon, lat, bag, crs = c("longlat", "daymet", "native")[1],
  bbox = FALSE)
}
\arguments{
\item{lon}{vector of longitudes}

\item{lat}{vector of latitudes}

\item{bag}{a list containing look up values of daymet array lons and lats}

\item{crs}{character indicating what proj the inputs belong to "longlat" or "daymet"}

\item{bbox}{logical if TRUE assume we are trying to find the enclosing bbox, which
trims the result to the maximum range in x and.}
}
\value{
a two column vector of closest indices in [x,y] order
}
\description{
Compute indices into daymet arrays from lon/lat or from the native
daymet Lambert conformal conic coordinates
}
