% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{daymet_nav}
\alias{daymet_nav}
\title{Prepare a navigation list for subsetting daymet arrays}
\usage{
daymet_nav(bb, layer = 1, bag, ...)
}
\arguments{
\item{bb}{a 4-element bounding box in [left, right, bottom, top] order}

\item{layer}{sequences of one or more contiguous layer numbers (in range of 1-365 or 1-12)}

\item{bag}{a list containing the lons, lats, crs and resolution of native daymet data}

\item{...}{further arguments for \code{\link{daymet_index}}}
}
\value{
a list with
 \itemize{
    \item{start a 3 element vector of starts for x, y and t}
    \item{count a 3 element vector of run lengths for x, y and t}
    \item{template an empty Raster layer object}
}
}
\description{
Prepare a navigation list for subsetting daymet arrays
}
