% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{daymet_raster}
\alias{daymet_raster}
\title{Retrieve one or more raster layers of daymet data}
\usage{
daymet_raster(x, bb = c(-77, -51.5, 37.9, 56.7), layers = 1,
  bag = daymet_bag(x), ...)
}
\arguments{
\item{x}{a ncdf4 object}

\item{bb}{numeric, 4 element bounding box specified in degrees lon and lat in
[left, right, bottom, top] order where west and south are negative.}

\item{layers}{numeric one or more 1-based layer index.  If more than one then a
multi-layer raster is returned.  Note that contiguous layers (e.g. 7, 8, 9) are faster
to retrieve than non-contiguous layers (e.g. 3, 7, 9).  The units of the layers
vary with the data source: for dailies layer is day-of-year (1-365).  For monthlies
layer is month-of-year (1-12).}

\item{bag}{list produced by \code{daymet_bag()}}

\item{...}{further arguments for \code{\link{daymet_index}}}
}
\value{
a Raster* or NULL
}
\description{
Retrieve one or more raster layers of daymet data
}
