% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{daymet_station_uri}
\alias{daymet_station_uri}
\title{Build a daymet uri for station data}
\usage{
daymet_station_uri(year = 2018, param = "tmin",
  baseuri = "https://thredds.daac.ornl.gov/thredds/dodsC/ornldaac/1391")
}
\arguments{
\item{year}{4 digit year (just one please)}

\item{param}{character the parameter name}

\item{baseuri}{character, the base uri}
}
\value{
charcater uri
}
\description{
Build a daymet uri for station data
}
