% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{daymet_version}
\alias{daymet_version}
\title{Retrieve the current dayment version identifier}
\usage{
daymet_version()
}
\value{
charcater version identifier
}
\description{
Retrieve the current dayment version identifier
}
