% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{DaymetGrid}
\alias{DaymetGrid}
\title{Establish a DaymetGridRefClass object}
\usage{
DaymetGrid(uri)
}
\arguments{
\item{uri}{the uri of the daymet grid resource}
}
\value{
DaymetGridRefRefClass object
}
\description{
Establish a DaymetGridRefClass object
}
