% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\docType{class}
\name{DaymetGridRefClass-class}
\alias{DaymetGridRefClass-class}
\alias{DaymetGridRefClass}
\title{An object that connects to opendap daymet grids using ncdf4}
\description{
An object that connects to opendap daymet grids using ncdf4
}
\section{Fields}{

\describe{
\item{\code{NC}}{the ncdf4 class objects}

\item{\code{bag}}{a list (bag of goodies) useful for navigating the ncdf object, see \code{\link{daymet_bag}}}
}}


