% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStations}
\alias{DaymetStations}
\title{Establish a DaymetStationsRefClass object}
\usage{
DaymetStations(uri = daymet_station_uri())
}
\arguments{
\item{uri}{the uri of the daymet stations resource}
}
\value{
DaymetStationsRefClass object
}
\description{
Establish a DaymetStationsRefClass object
}
