% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\docType{class}
\name{DaymetStationsRefClass-class}
\alias{DaymetStationsRefClass-class}
\alias{DaymetStationsRefClass}
\title{An object that connects to opendap daymet stations using ncdf4}
\description{
An object that connects to opendap daymet stations using ncdf4
}
\section{Fields}{

\describe{
\item{\code{NC}}{the ncdf4 class objects}

\item{\code{stns}}{numeric, the number of stations represented}

\item{\code{.station_id}}{private, please use method station_id()}

\item{\code{.station_name}}{private, please use method station_name()}

\item{\code{.stnz}}{private, please use method stnz()}

\item{\code{.stnx}}{private, please use method stnx()}

\item{\code{.stny}}{private, please use method stny()}

\item{\code{.stn_lon}}{private, please use method stn_lon()}

\item{\code{.stny_lat}}{private, please use method stn_lat()}

\item{\code{.dates}}{private, please use method dates()}
}}


