% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_closest_stations}
\alias{DaymetStationsRefClass_closest_stations}
\title{Find the n closest stations to the lon/lat coordinates provided.}
\arguments{
\item{lon}{numeric, the longitude locations}

\item{lat}{numeric, the latitude locations}

\item{n}{numeric, the number of stations to retrieve}

\item{longlat}{logical, if TRUE then find great circle distance. See
\code{sp::spDistN1()}}
}
\value{
tibble of the n closest stations with the following variables
 \itemize{
 \item{index the index of the station (order with station_id, station_name, etc)}
 \item{dist numeric, the distance in kilometers from station to requested point}
 }
}
\description{
Find the n closest stations to the lon/lat coordinates provided.
}
