% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_station_data}
\alias{DaymetStationsRefClass_station_data}
\title{Get station data by station name, station id or index within resource.}
\arguments{
\item{name}{character the station name}

\item{id}{character, if name is not provided then we try to match the station_id provided}

\item{index}{numeric, if name and id are not provided, then we use the index provided}
}
\value{
tibble of zero or more rows of variables date, obs and pred data values
}
\description{
At least one of name, id or index must be provided.
}
