% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_station_id}
\alias{DaymetStationsRefClass_station_id}
\title{Retrieve the station ids}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
character vector of station ids
}
\description{
Retrieve the station ids
}
