% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_station_name}
\alias{DaymetStationsRefClass_station_name}
\title{Retrieve the station names}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
character vector of station names
}
\description{
Retrieve the station names
}
