% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_stn_lat}
\alias{DaymetStationsRefClass_stn_lat}
\title{Retrieve the station lat locations}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
numeric vector of lat locations
}
\description{
Retrieve the station lat locations
}
