% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_stn_lon}
\alias{DaymetStationsRefClass_stn_lon}
\title{Retrieve the station lon locations}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
numeric vector of lon locations
}
\description{
Retrieve the station lon locations
}
