% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_stnx}
\alias{DaymetStationsRefClass_stnx}
\title{Retrieve the station x locations}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
numeric vector of x locations
}
\description{
Retrieve the station x locations
}
