% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_stny}
\alias{DaymetStationsRefClass_stny}
\title{Retrieve the station y locations}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
numeric vector of y locations
}
\description{
Retrieve the station y locations
}
