% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{DaymetStationsRefClass_stnz}
\alias{DaymetStationsRefClass_stnz}
\title{Retrieve the station z elevations}
\arguments{
\item{index}{numeric, the indices to return}
}
\value{
numeric vector of z elevations
}
\description{
Retrieve the station z elevations
}
