% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{fetch_daymet_raster}
\alias{fetch_daymet_raster}
\title{Retrieve one or more raster layers of daymet data}
\usage{
fetch_daymet_raster(x, param, nav)
}
\arguments{
\item{x}{a ncdf4 object}

\item{param}{character, name of the variable}

\item{nav}{list of navigation items, see \code{\link{daymet_nav}}}
}
\value{
RasterLayer or RasterStack
}
\description{
Retrieve one or more raster layers of daymet data
}
