% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_crs}
\alias{get_crs}
\title{Retrieve a coordiante reference string by name}
\usage{
get_crs(name = c("longlat", "lcc", "native")[1])
}
\arguments{
\item{name}{character - either 'longlat', or 'daymet' or 'native' The latter
is the same as 'lcc'.}
}
\value{
coordinate reference string in proj4 style
}
\description{
Retrieve a coordiante reference string by name
}
