% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncss.R
\name{ncss_basename}
\alias{ncss_basename}
\title{Extract the file basename out of a ncss_url}
\usage{
ncss_basename(uri = ncss_base_url())
}
\arguments{
\item{uri}{character, the uri to parse}
}
\value{
character, the file basename
}
\description{
Extract the file basename out of a ncss_url
}
