% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncss.R
\name{ncss_load_brick}
\alias{ncss_load_brick}
\title{Load NCSS data as a raster brick}
\usage{
ncss_load_brick(filename, modify_spatial = TRUE)
}
\arguments{
\item{filename}{character, the path and name of the NCDF file to read}

\item{modify_spatial}{logical, if TRUE then update the projection and resolution}
}
\value{
\{code{\link[raster]{brick}} object
}
\description{
Load NCSS data as a raster brick
}
