% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncss.R
\name{ncss_query_misc}
\alias{ncss_query_misc}
\title{Handle miscellaneous query elements}
\usage{
ncss_query_misc(
  disableProjSubset = "on",
  horizStride = 1,
  timeStride = 1,
  accept = "netcdf"
)
}
\arguments{
\item{disableProjSubset}{character either on or off}

\item{horizStride}{numeric}

\item{timeStride}{numeric}

\item{accept}{charcater}
}
\description{
Handle miscellaneous query elements
}
