% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncss.R
\name{ncss_url}
\alias{ncss_url}
\title{Generate a NCSS url with subset qualifiers}
\usage{
ncss_url(
  var = "dayl",
  region = "na",
  dates = as.Date(c("2019-01-01", "2019-01-31")),
  bb = c(-157.0759, -6.1376, 6.0761, 83.0163),
  ...
)
}
\arguments{
\item{var}{character, the variable name}

\item{region}{character, the name of the region (default 'na')}

\item{dates}{Date 2 element Date vector with inclusive start/stop dates.
These should fall within one calendar year.}

\item{bb}{a 4 element bounding box [west, east, south, north]}

\item{...}{items for \code{\link{ncss_query_misc}}}
}
\value{
charcater URL with query terms
}
\description{
Generate a NCSS url with subset qualifiers
}
\seealso{
\url{https://daymet.ornl.gov/web_services}
}
