% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{param_from_uri}
\alias{param_from_uri}
\title{Guess the parameter from the uri}
\usage{
param_from_uri(uri = c("daymet_v3_tmin_2018_na.nc4"))
}
\arguments{
\item{uri}{character the uri to the resource}
}
\value{
character best guess of the parameter short hand
}
\description{
Guess the parameter from the uri
}
