% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{param_from_uri_grid}
\alias{param_from_uri_grid}
\title{Guess the parameter from the uri for gridded data}
\usage{
param_from_uri_grid(uri = c("daymet_v4_tmin_2018_na.nc4"))
}
\arguments{
\item{uri}{character the uri to the resource}
}
\value{
character best guess of the parameter short hand
}
\description{
Guess the parameter from the uri for gridded data
}
