% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{param_from_uri_station}
\alias{param_from_uri_station}
\title{Guess the parameter from the uri for station data}
\usage{
param_from_uri_station(uri = c("daymet_v4_stnsxval_prcp_na_1980.nc4"))
}
\arguments{
\item{uri}{character the uri to the resource}
}
\value{
character best guess of the parameter short hand
}
\description{
Guess the parameter from the uri for station data
}
