% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{points_transfrom}
\alias{points_transfrom}
\title{Transform a set of coordinates}
\usage{
points_transfrom(x, y = NULL, from_proj = get_crs("longlat"),
  to_proj = get_crs("daymet"))
}
\arguments{
\item{x}{numeric locations such as decimal degrees longitude.  If y is not provided then
x must be a two column matrix providing [x,y].}

\item{y}{numeric locations such as decimal degrees latitude}

\item{from_proj}{projection of the input coordinates}

\item{to_proj}{projection of the output coordinates}
}
\value{
2 column matrix of lcc [x,y]
}
\description{
Transform a set of coordinates
}
